removed AbstractFacetPartitionSpec, finished reworking FacetedPartitioner and related factory
