+++ connector/src/main/java/io/github/dejarol/azure/search/spark/connector/core/config/ConfigException.java
@@ -3,6 +3,7 @@ package io.github.dejarol.azure.search.spark.connector.core.config;
 import io.github.dejarol.azure.search.spark.connector.core.Constants;
 import org.apache.spark.sql.DataFrameReader;
 import org.apache.spark.sql.DataFrameWriter;
+import org.jetbrains.annotations.Contract;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -80,6 +81,7 @@ public class ConfigException
      * @param key configuration key
      * @param value configuration value
      * @param cause exception cause
+     * @return an instance
      */
 
    public static @NotNull ConfigException forIllegalOptionValue(
@@ -101,6 +103,7 @@ public class ConfigException
      * @param key configuration key
      * @param value configuration value
      * @param reason reason
+     * @return an instance
      */
 
     public static @NotNull ConfigException forIllegalOptionValue(
@@ -115,5 +118,31 @@ public class ConfigException
         );
 
         return new ConfigException(supplier, null);
+    }
+
+    /**
+     * Create an instance for an option with an illegal value, usually generated by a configuration
+     * whose raw value could not be converted to a different target type
+     * @param key configuration key
+     * @param value raw configuration value
+     * @param reason reason
+     * @param cause cause
+     * @return an instance
+     */
+
+    @Contract("_, _, _, _ -> new")
+    public static @NotNull ConfigException forIllegalOptionValue(
+            @NotNull String key,
+            @NotNull String value,
+            @NotNull String reason,
+            @NotNull Throwable cause
+    ) {
+
+        Supplier<String> supplier = () -> String.format(
+                "Illegal value for option '%s' (%s). Reason: %s",
+                key, value, reason
+        );
+
+        return new ConfigException(supplier, cause);
     }
 }
