+++ connector/src/main/scala/io/github/dejarol/azure/search/spark/connector/core/schema/SearchFieldDataTypeOperations.scala
@@ -75,24 +75,6 @@ class SearchFieldDataTypeOperations(private val input: SearchFieldDataType)
    */
 
   final def isGeoPoint: Boolean = input.equals(SearchFieldDataType.GEOGRAPHY_POINT)
-
-  /**
-   * Evaluate if this type is a candidate for partitioning.
-   * <br>
-   * Only numeric or date time types are candidates
-   * @return true for numeric or date time types
-   */
-
-  final def isCandidateForPartitioning: Boolean = isDateTime || isNumeric
-
-  /**
-   * Evaluate if this type is a candidate for faceting
-   * <br>
-   * Only string or numeric types are good candidates
-   * @return true for string or numeric types
-   */
-
-  final def isCandidateForFaceting: Boolean = isString || isNumeric
 }
 
 private object SearchFieldDataTypeOperations {
