+++ connector/src/main/scala/io/github/dejarol/azure/search/spark/connector/read/partitioning/AbstractFacetPartition.scala
@@ -1,13 +1,8 @@
 package io.github.dejarol.azure.search.spark.connector.read.partitioning
 
-import com.azure.search.documents.indexes.models.SearchField
-import io.github.dejarol.azure.search.spark.connector.core.schema._
-import io.github.dejarol.azure.search.spark.connector.core.utils.StringUtils
-
 /**
  * Parent class for partitions created by a [[FacetedPartitioner]] by retrieving a set of values
  * from a Search field that is facetable and filterable
- *
  * @param partitionId partition id
  * @param facetFieldName name of the field used for faceting values
  */
@@ -17,71 +12,4 @@ abstract class AbstractFacetPartition(
                                        protected val facetFieldName: String
                                      )
   extends AbstractSearchPartition(partitionId) {
-}
-
-object AbstractFacetPartition {
-
-  /**
-   * Trait for formatting facet values to string
-   * <br>
-   * Formatted values will be used for filtering documents within partitions
-   */
-
-  private sealed trait FacetToStringFunction extends (Any => String)
-
-  /**
-   * Generate a set of partitions from values retrieved from a facetable field
-   * @param facetField facet field
-   * @param facets facet field values
-   * @return a collection of Search partitions
-   */
-
-  def createCollection(
-                        facetField: SearchField,
-                        facets: Seq[Any]
-                      ): Seq[AbstractFacetPartition] = {
-
-    val toStringFunction = getFunction(facetField)
-    val facetFieldName: String = facetField.getName
-    val facetStringValues: Seq[String] = facets.map(toStringFunction)
-
-    // Create as many partitions as the number of retrieved facet values
-    val partitionsForFacetsValues: Seq[AbstractFacetPartition] = facetStringValues
-      .zipWithIndex.map {
-        case (value, partitionId) =>
-          FacetValuePartition(partitionId, facetFieldName, value)
-      }
-
-    // Add another partition for either null values or other facet values
-    val partitionForEitherNullOrOtherFacetValues = FacetNullValuePartition(
-      facetFieldName,
-      facetStringValues
-    )
-
-    partitionsForFacetsValues :+ partitionForEitherNullOrOtherFacetValues
-  }
-
-  /**
-   * Get the function corresponding to a Search type
-   * @param searchField facetable field
-   * @throws IllegalStateException for Search types (should not occur)
-   * @return a function for value formatting
-   */
-
-  @throws[IllegalStateException]
-  private def getFunction(searchField: SearchField): FacetToStringFunction = {
-
-    val searchFieldDataType = searchField.getType
-    if (searchFieldDataType.isString) {
-      new FacetToStringFunction {
-        override def apply(v1: Any): String = StringUtils.singleQuoted(v1.asInstanceOf[String])
-      }
-    } else if (searchFieldDataType.isNumeric) {
-      new FacetToStringFunction {
-        override def apply(v1: Any): String = String.valueOf(v1)
-      }
-    } else {
-      throw new IllegalStateException(f"No facet to string function defined for $searchField")
-    }
-  }
 }
