+++ connector/src/test/scala/io/github/dejarol/azure/search/spark/connector/read/partitioning/FacetedPartitionerSpec.scala
@@ -1,112 +1,31 @@
 package io.github.dejarol.azure.search.spark.connector.read.partitioning
 
-import com.azure.search.documents.indexes.models.{SearchField, SearchFieldDataType}
-import io.github.dejarol.azure.search.spark.connector.core.schema.SearchFieldFeature
+import io.github.dejarol.azure.search.spark.connector.core.JavaScalaConverters
 import io.github.dejarol.azure.search.spark.connector.{BasicSpec, FieldFactory}
-import org.scalatest.EitherValues
 
 class FacetedPartitionerSpec
   extends BasicSpec
-    with FieldFactory
-      with EitherValues {
+    with FieldFactory {
 
-  private lazy val first = "first"
-
-  /**
-   * Enable features on a [[SearchField]]
-   * @param field    field
-   * @param features features to enable
-   */
-
-  private def enableFeatures(
-                              field: SearchField,
-                              features: SearchFieldFeature*
-                            ): SearchField = {
-
-    features.foldLeft(field) {
-      case (field, feature) =>
-        feature.enableOnField(field)
-    }
-  }
-
-  describe(`object`[FacetedPartitioner]) {
+  describe(anInstanceOf[FacetedPartitioner]) {
     describe(SHOULD) {
-      describe("evaluate the number of partitions returning") {
-        describe("a Right when") {
-          it("no number is provided") {
+      it("create a collection of partitions") {
 
-            FacetedPartitioner.evaluatePartitionNumber(
-              None
-            ) shouldBe 'right
-          }
+        val facetFieldName = "field"
+        val facetValues = Seq("v1", "v2", "v3")
+        val actual = JavaScalaConverters.listToSeq(
+          FacetedPartitioner(facetFieldName, facetValues)
+            .createPartitions()
+        )
 
-          it("a valid number is provided") {
+        actual should have size(facetValues.size + 1)
+        actual.count {
+          _.isInstanceOf[FacetValuePartition]
+        } shouldBe facetValues.size
 
-            FacetedPartitioner.evaluatePartitionNumber(
-              Some(2)
-            ) shouldBe 'right
-          }
-        }
-
-        describe("a Left for") {
-          it("invalid numbers") {
-
-            FacetedPartitioner.evaluatePartitionNumber(
-              Some(1)
-            ) shouldBe 'left
-          }
-        }
-      }
-
-      describe("evaluate if a Search field is eligible for faceting") {
-        describe("returning a Right for") {
-          it("an existing facetable and filterable field") {
-
-            FacetedPartitioner.getCandidateFacetField(
-              first,
-              Seq(
-                enableFeatures(
-                  createSearchField(first, SearchFieldDataType.STRING),
-                  SearchFieldFeature.FACETABLE,
-                  SearchFieldFeature.FILTERABLE
-                )
-              )
-            ) shouldBe 'right
-          }
-        }
-
-        describe("a Left for") {
-          it("non-existing field") {
-
-            FacetedPartitioner.getCandidateFacetField(
-              first,
-              Seq.empty
-            ) shouldBe 'left
-          }
-
-          it("a non-filterable or facetable field") {
-
-            FacetedPartitioner.getCandidateFacetField(
-              first,
-              Seq(
-                enableFeatures(
-                  createSearchField(first, SearchFieldDataType.STRING),
-                  SearchFieldFeature.FACETABLE
-                )
-              )
-            ) shouldBe 'left
-
-            FacetedPartitioner.getCandidateFacetField(
-              first,
-              Seq(
-                enableFeatures(
-                  createSearchField(first, SearchFieldDataType.STRING),
-                  SearchFieldFeature.FILTERABLE
-                )
-              )
-            ) shouldBe 'left
-          }
-        }
+        actual.count {
+          _.isInstanceOf[FacetNullValuePartition]
+        } shouldBe 1
       }
     }
   }
